refactor(desktop): one stage installer for both web shells, and main.rs read in pieces
The desktop and browser shells carried the same installStage body once the
desktop one stopped painting transparent; it now lives with fitStage in
@kroma/tv/stage and both shells import it. The Linux window-event and setup
closures leave main() for named functions, and the GPU boot probe leaves
apply_env for its own, so neither reads as one long branch.
