+++ clients/desktop/src/main.tsx
@@ -1,5 +1,6 @@
 import 'virtual:kroma-tv.css';
 import { mountTv } from '@kroma/tv/mount';
+import { installStage } from '@kroma/tv/stage';
 // Display-matched grade of the brand-intro film, bundled by THIS shell only:
 // 1080p60 is a quarter of the shared 4K60 film's decode work and
 // indistinguishable in a desktop window, and H.264 (not HEVC like the shared
@@ -9,7 +10,6 @@ import { mountTv } from '@kroma/tv/mount';
 // KromaIntro/constants.ts.
 import introFilm from './assets/kroma-intro-h264-1080.mp4';
 import { startGamepadBridge } from './gamepad';
-import { installStage } from './stage';
 import { startUpdater } from './updater';
 import { installVideoHole, nativePlane } from './video-hole';
 
@@ -17,7 +17,7 @@ import { installVideoHole, nativePlane } from './video-hole';
 // in fixed pixels against that canvas (PosterGrid's 8 x 203px columns, the nav row,
 // the episode column), so a free-size window narrower than 1920 does not shrink the
 // layout, it clips it. Fitted the same way as the Steam Deck panel and the browser
-// shell (see ./stage and clients/tv-web/src/stage.ts). A genuinely fluid 10-foot
+// shell (`installStage`, @kroma/tv/stage). A genuinely fluid 10-foot
 // layout is a design-system change, not a shell one.
 const fixedScreen = nativePlane();
 installStage();
